Refactor, move columntype and add supports to interface
